Don't create duplicated pastes
A lot of users will click the Save this button without changing
anything from the default code example. This was recreating the same
default paste over and over again. This will first try and find a
duplicate paste and redirect to that. If none exists, then it will
create a new paste.